move comparative widgets to their own folder
